% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{cal_abs}
\alias{cal_abs}
\title{cal_abs}
\usage{
cal_abs(lipid_tag, lipid_table, is_tag, is_table, match_item)
}
\arguments{
\item{lipid_tag}{lipid_tag}

\item{lipid_table}{lipid_table}

\item{is_tag}{is_tag}

\item{is_table}{is_table}

\item{match_item}{match_item}
}
\description{
cal_abs
}
\author{
Xiaotao Shen
}
