% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{clean_is_table}
\alias{clean_is_table}
\title{clean_is_table}
\usage{
clean_is_table(x, polarity = c("positive", "negative"))
}
\arguments{
\item{x}{x}

\item{polarity}{polarity}
}
\description{
clean_is_table
}
\author{
Xiaotao Shen
}
