% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{combine_pos_neg_quantification}
\alias{combine_pos_neg_quantification}
\title{combine_pos_neg_quantification}
\usage{
combine_pos_neg_quantification(
  path = ".",
  express_data_abs_ug_ml_pos,
  express_data_abs_um_pos,
  variable_info_abs_pos,
  express_data_abs_ug_ml_neg,
  express_data_abs_um_neg,
  variable_info_abs_neg
)
}
\arguments{
\item{path}{path}

\item{express_data_abs_ug_ml_pos}{express_data_abs_ug_ml_pos}

\item{express_data_abs_um_pos}{express_data_abs_um_pos}

\item{variable_info_abs_pos}{variable_info_abs_pos}

\item{express_data_abs_ug_ml_neg}{express_data_abs_ug_ml_neg}

\item{express_data_abs_um_neg}{express_data_abs_um_neg}

\item{variable_info_abs_neg}{variable_info_abs_neg}
}
\description{
combine_pos_neg_quantification
}
\author{
Xiaotao Shen
}
