% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{combine_quantification_data_and_feature_info}
\alias{combine_quantification_data_and_feature_info}
\title{combine_quantification_data_and_feature_info}
\usage{
combine_quantification_data_and_feature_info(
  quantification_data,
  feature_info,
  targeted_table_type = c("is", "lipid")
)
}
\arguments{
\item{quantification_data}{quantification_data}

\item{feature_info}{feature_info}

\item{targeted_table_type}{targeted_table_type}
}
\description{
combine_quantification_data_and_feature_info
}
\author{
Xiaotao Shen
}
