% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{extract_targeted_peaks}
\alias{extract_targeted_peaks}
\title{extract_targeted_peaks}
\usage{
extract_targeted_peaks(
  path = ".",
  output_path_name = "Result",
  targeted_targeted_peak_table_name = "is_table.xlsx",
  forced_targeted_peak_table_name = "forced_table.xlsx",
  from_lipid_search = FALSE,
  sample_numer_show = 5,
  fit.gaussian = TRUE,
  integrate_xcms = FALSE,
  output_eic = TRUE,
  output_integrate = TRUE,
  ppm = 25,
  rt.tolerance = 90,
  threads = 3,
  facet = TRUE
)
}
\arguments{
\item{path}{path}

\item{output_path_name}{output_path_name}

\item{targeted_targeted_peak_table_name}{targeted_targeted_peak_table_name}

\item{forced_targeted_peak_table_name}{forced_targeted_peak_table_name}

\item{from_lipid_search}{from_lipid_search}

\item{sample_numer_show}{sample_numer_show}

\item{fit.gaussian}{fit.gaussian}

\item{integrate_xcms}{integrate_xcms}

\item{output_eic}{output_eic}

\item{output_integrate}{output_integrate}

\item{ppm}{ppm}

\item{rt.tolerance}{rt.tolerance}

\item{threads}{threads}

\item{facet}{facet}
}
\description{
extract_targeted_peaks
}
\author{
Xiaotao Shen
}
