% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{extractPeaks2}
\alias{extractPeaks2}
\title{extractPeaks2}
\usage{
extractPeaks2(
  path = ".",
  output_path_name = "Result",
  ppm = 15,
  threads = 4,
  is.table = "is.xlsx",
  mz = NULL,
  rt = NULL,
  rt.tolerance = 40,
  msLevel = 1,
  filled = TRUE,
  include = c("apex_within", "any", "none")
)
}
\arguments{
\item{path}{path}

\item{output_path_name}{output_path_name}

\item{ppm}{ppm}

\item{threads}{threads}

\item{is.table}{is.table}

\item{mz}{mz}

\item{rt}{rt}

\item{rt.tolerance}{rt.tolerance}

\item{msLevel}{msLevel}

\item{filled}{filled}

\item{include}{include}
}
\description{
extractPeaks2
}
\author{
Xiaotao Shen
}
