% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{from_quantification_table_to_rt_table}
\alias{from_quantification_table_to_rt_table}
\title{from_quantification_table_to_rt_table}
\usage{
from_quantification_table_to_rt_table(path = ".")
}
\arguments{
\item{path}{path}
}
\description{
from_quantification_table_to_rt_table
}
\author{
Xiaotao Shen
}
