% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{from_rt_table_to_is_info}
\alias{from_rt_table_to_is_info}
\title{from_rt_table_to_is_info}
\usage{
from_rt_table_to_is_info(
  polarity = c("positive", "negative"),
  is.info,
  path = "."
)
}
\arguments{
\item{polarity}{polarity}

\item{is.info}{is.info}

\item{path}{path}
}
\description{
from_rt_table_to_is_info
}
\author{
Xiaotao Shen
}
