% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_absolute_quantification.R
\name{get_absolute_quantification}
\alias{get_absolute_quantification}
\title{get_absolute_quantification}
\usage{
get_absolute_quantification(
  path = ".",
  is_quantification_table,
  lipid_quantification_table,
  sample_info,
  match_item
)
}
\arguments{
\item{path}{path.}

\item{is_quantification_table}{is_quantification_table}

\item{lipid_quantification_table}{lipid_quantification_table}

\item{sample_info}{sample_info}

\item{match_item}{match_item}
}
\value{
Peak plot for each internal standard.
}
\description{
get_absolute_quantification
}
\author{
Xiaotao Shen
}
