% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_is_quantify_table}
\alias{get_is_quantify_table}
\title{get_is_quantify_table}
\usage{
get_is_quantify_table(
  is_table,
  peak_table,
  mz.tol = 25,
  rt.tol = 90,
  figure = FALSE,
  polarity = c("positive", "negative")
)
}
\arguments{
\item{is_table}{is_table}

\item{peak_table}{peak_table}

\item{mz.tol}{mz.tol}

\item{rt.tol}{rt.tol}

\item{figure}{figure}

\item{polarity}{polarity}
}
\description{
get_is_quantify_table
}
\author{
Xiaotao Shen
}
