% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_IS_RT2.R
\name{get_IS_RT}
\alias{get_IS_RT}
\title{get_IS_RT}
\usage{
get_IS_RT(
  path = ".",
  is_info_table,
  polarity = c("positive", "negative"),
  threads = 3,
  rerun = FALSE
)
}
\arguments{
\item{path}{path.}

\item{is_info_table}{is_info_table}

\item{polarity}{polarity}

\item{threads}{threads}

\item{rerun}{rerun}
}
\description{
get_IS_RT
}
\author{
Xiaotao Shen
}
