% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lipid_absolute_quantification.R
\name{get_lipid_absolute_quantification}
\alias{get_lipid_absolute_quantification}
\title{get_lipid_absolute_quantification}
\usage{
get_lipid_absolute_quantification(
  path = ".",
  is_info_name_pos = "IS_information.xlsx",
  is_info_name_neg = "IS_information.xlsx",
  use_manual_is_info = FALSE,
  lipid_annotation_table_pos = "lipid_annotation_table_pos.xlsx",
  lipid_annotation_table_neg = "lipid_annotation_table_neg.xlsx",
  chol_rt = 1169,
  output_integrate = TRUE,
  forced_targeted_peak_table_name = NULL,
  fit.gaussian = TRUE,
  integrate_xcms = TRUE,
  output_eic = TRUE,
  ppm = 40,
  rt.tolerance = 180,
  threads = 3,
  rerun = FALSE,
  which_group_for_rt_confirm = "QC",
  match_item_pos = list(Cer = "d18:1 (d7)-15:0 Cer", ChE = c("18:1(d7) Chol Ester",
    "Cholesterol (d7)"), Chol = "Cholesterol (d7)", DG = "15:0-18:1(d7) DAG", LPC =
    "18:1(d7) Lyso PC", LPE = "18:1(d7) Lyso PE", MG = "18:1 (d7) MG", PA =
    "15:0-18:1(d7) PA (Na Salt)", PC = "15:0-18:1(d7) PC", PE = "15:0-18:1(d7) PE", PG =
    "15:0-18:1(d7) PG (Na Salt)", PI = "15:0-18:1(d7) PI (NH4 Salt)", PPE =
    "C18(Plasm)-18:1(d9) PE", PS = "15:0-18:1(d7) PS (Na Salt)", SM =
    "d18:1-18:1(d9) SM", TG = "15:0-18:1(d7)-15:0 TAG"),
  match_item_neg = list(Cer = "d18:1 (d7)-15:0 Cer", Chol = "Cholesterol (d7)", ChE =
    c("18:1(d7) Chol Ester", "Cholesterol (d7)"), LPC = "18:1(d7) Lyso PC", LPE =
    "18:1(d7) Lyso PE", PC = "15:0-18:1(d7) PC", PE = "15:0-18:1(d7) PE", PG =
    "15:0-18:1(d7) PG (Na Salt)", PI = "15:0-18:1(d7) PI (NH4 Salt)", PPE =
    "C18(Plasm)-18:1(d9) PE", PS = "15:0-18:1(d7) PS (Na Salt)", SM =
    "d18:1-18:1(d9) SM")
)
}
\arguments{
\item{path}{path.}

\item{is_info_name_pos}{is_info_name_pos}

\item{is_info_name_neg}{is_info_name_neg}

\item{use_manual_is_info}{use_manual_is_info}

\item{lipid_annotation_table_pos}{lipid_annotation_table_pos}

\item{lipid_annotation_table_neg}{lipid_annotation_table_neg}

\item{chol_rt}{chol_rt}

\item{output_integrate}{output_integrate}

\item{forced_targeted_peak_table_name}{forced_targeted_peak_table_name}

\item{fit.gaussian}{fit.gaussian}

\item{integrate_xcms}{integrate_xcms}

\item{output_eic}{output_eic}

\item{ppm}{ppm}

\item{rt.tolerance}{rt.tolerance}

\item{threads}{threads}

\item{rerun}{rerun}

\item{which_group_for_rt_confirm}{which_group_for_rt_confirm}

\item{match_item_pos}{match_item_pos}

\item{match_item_neg}{match_item_neg}
}
\value{
Peak plot for each internal standard.
}
\description{
get_lipid_absolute_quantification
}
\author{
Xiaotao Shen
}
