% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_quantification_data2}
\alias{get_quantification_data2}
\title{get_quantification_data2}
\usage{
get_quantification_data2(
  path = ".",
  is_tag,
  is_table,
  lipid_tag,
  lipid_table,
  match_item
)
}
\arguments{
\item{path}{path}

\item{is_tag}{is_tag}

\item{is_table}{is_table}

\item{lipid_tag}{lipid_tag}

\item{lipid_table}{lipid_table}

\item{match_item}{match_item}
}
\description{
get_quantification_data2
}
\author{
Xiaotao Shen
}
