% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_relative_quantification2.R
\name{get_relative_quantification}
\alias{get_relative_quantification}
\title{get_relative_quantification}
\usage{
get_relative_quantification(
  path = ".",
  output_path_name = "Result",
  targeted_table_name,
  sample_info,
  targeted_table_type = c("is", "lipid"),
  polarity = c("positive", "negative"),
  chol_rt = 1169,
  output_integrate = TRUE,
  forced_targeted_peak_table_name = NULL,
  fit.gaussian = TRUE,
  integrate_xcms = TRUE,
  output_eic = TRUE,
  ppm = 40,
  rt.tolerance = 180,
  threads = 3,
  rerun = FALSE
)
}
\arguments{
\item{path}{path.}

\item{output_path_name}{output_path_name}

\item{targeted_table_name}{targeted_table_name}

\item{sample_info}{sample_info}

\item{targeted_table_type}{targeted_table_type}

\item{polarity}{polarity}

\item{chol_rt}{chol_rt}

\item{output_integrate}{output_integrate}

\item{forced_targeted_peak_table_name}{forced_targeted_peak_table_name}

\item{fit.gaussian}{fit.gaussian}

\item{integrate_xcms}{integrate_xcms}

\item{output_eic}{output_eic}

\item{ppm}{ppm}

\item{rt.tolerance}{rt.tolerance}

\item{threads}{threads}

\item{rerun}{rerun}
}
\description{
get_IS_RT
}
\author{
Xiaotao Shen
}
