% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lipidflow_logo.R
\name{lipidflow_logo}
\alias{lipidflow_logo}
\title{Show the base information of lipidflow package}
\usage{
lipidflow_logo()
}
\value{
A ASCII log of lipidflow
}
\description{
Show the base information of lipidflow package.
}
\examples{
lipidflow_logo()
}
\author{
Xiaotao Shen
\email{shenxt@stanford.edu}
}
