% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attach.R
\name{lipidflow_packages}
\alias{lipidflow_packages}
\title{List all packages in the lipidflow}
\usage{
lipidflow_packages(include_self = TRUE)
}
\arguments{
\item{include_self}{Include lipidflow in the list?}
}
\description{
List all packages in the lipidflow
}
\examples{
lipidflow_packages()
}
