% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{match_sample_peak_table_lipid_search}
\alias{match_sample_peak_table_lipid_search}
\title{match_sample_peak_table_lipid_search}
\usage{
match_sample_peak_table_lipid_search(peak_table, lipid_search)
}
\arguments{
\item{peak_table}{peak_table}

\item{lipid_search}{lipid_search}
}
\description{
match_sample_peak_table_lipid_search
}
\author{
Xiaotao Shen
}
