% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organize_results.R
\name{organize_result}
\alias{organize_result}
\title{organize result}
\usage{
organize_result(
  path = ".",
  match_item_pos = list(Cer = "d18:1 (d7)-15:0 Cer", ChE = c("18:1(d7) Chol Ester",
    "Cholesterol (d7)"), Chol = "Cholesterol (d7)", DG = "15:0-18:1(d7) DAG", LPC =
    "18:1(d7) Lyso PC", LPE = "18:1(d7) Lyso PE", MG = "18:1 (d7) MG", PA =
    "15:0-18:1(d7) PA (Na Salt)", PC = "15:0-18:1(d7) PC", PE = "15:0-18:1(d7) PE", PG =
    "15:0-18:1(d7) PG (Na Salt)", PI = "15:0-18:1(d7) PI (NH4 Salt)", PPE =
    "C18(Plasm)-18:1(d9) PE", PS = "15:0-18:1(d7) PS (Na Salt)", SM =
    "d18:1-18:1(d9) SM", TG = "15:0-18:1(d7)-15:0 TAG"),
  match_item_neg = list(Cer = "d18:1 (d7)-15:0 Cer", Chol = "Cholesterol (d7)", ChE =
    c("18:1(d7) Chol Ester", "Cholesterol (d7)"), LPC = "18:1(d7) Lyso PC", LPE =
    "18:1(d7) Lyso PE", PC = "15:0-18:1(d7) PC", PE = "15:0-18:1(d7) PE", PG =
    "15:0-18:1(d7) PG (Na Salt)", PI = "15:0-18:1(d7) PI (NH4 Salt)", PPE =
    "C18(Plasm)-18:1(d9) PE", PS = "15:0-18:1(d7) PS (Na Salt)", SM =
    "d18:1-18:1(d9) SM")
)
}
\arguments{
\item{path}{path.}

\item{match_item_pos}{match_item_pos}

\item{match_item_neg}{match_item_neg}
}
\value{
Peak plot for each internal standard.
}
\description{
organize result
}
\author{
Xiaotao Shen
}
