% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_result.R
\name{output_result}
\alias{output_result}
\title{output_result}
\usage{
output_result(path = ".", match_item_pos, match_item_neg)
}
\arguments{
\item{path}{work directory.}

\item{match_item_pos}{match_item_pos}

\item{match_item_neg}{match_item_neg}
}
\description{
output_result
}
\author{
Xiaotao Shen
}
