% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorganize_peak_plot.R
\name{reorganize_peak_plot}
\alias{reorganize_peak_plot}
\title{reorganize_peak_plot}
\usage{
reorganize_peak_plot(
  path = ".",
  plot_dir = "peak_shape",
  absolute_table,
  match_item
)
}
\arguments{
\item{path}{path.}

\item{plot_dir}{plot_dir}

\item{absolute_table}{absolute_table}

\item{match_item}{match_item}
}
\value{
Peak plot for each internal standard.
}
\description{
reorganize_peak_plot
}
\author{
Xiaotao Shen
}
