% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{showPeak2}
\alias{showPeak2}
\title{showPeak2}
\usage{
showPeak2(
  object,
  peak.index = 1,
  title.size = 12,
  lab.size = 12,
  axis.text.size = 12,
  alpha = 0.5,
  title = "",
  interactive = FALSE,
  area_data,
  raw_info,
  facet = TRUE
)
}
\arguments{
\item{object}{object}

\item{peak.index}{peak.index}

\item{title.size}{title.size}

\item{lab.size}{lab.size}

\item{axis.text.size}{axis.text.size}

\item{alpha}{alpha}

\item{title}{title}

\item{interactive}{interactive}

\item{area_data}{area_data}

\item{raw_info}{raw_info}

\item{facet}{facet}
}
\description{
showPeak2
}
\author{
Xiaotao Shen
}
