% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{tidy_lipidsearch_data}
\alias{tidy_lipidsearch_data}
\title{tidy_lipidsearch_data}
\usage{
tidy_lipidsearch_data(
  file,
  polarity = c("positive", "negative"),
  from = "lipidsearch"
)
}
\arguments{
\item{file}{file}

\item{polarity}{polarity}

\item{from}{from}
}
\description{
tidy_lipidsearch_data
}
\author{
Xiaotao Shen
}
