% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{trans_is_table}
\alias{trans_is_table}
\title{trans_is_table}
\usage{
trans_is_table(is_table, polarity = c("positive", "negative"), path = ".")
}
\arguments{
\item{is_table}{is_table}

\item{polarity}{polarity}

\item{path}{path}
}
\description{
trans_is_table
}
\author{
Xiaotao Shen
}
