% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_IS_RT.R
\name{get_IS_RT}
\alias{get_IS_RT}
\title{get_IS_RT}
\usage{
get_IS_RT(
  path = ".",
  is_info_table,
  polarity = c("positive", "negative"),
  threads = 3,
  rerun = FALSE,
  output_eic = TRUE
)
}
\arguments{
\item{path}{path.}

\item{is_info_table}{is_info_table}

\item{polarity}{polarity}

\item{threads}{threads}

\item{rerun}{rerun}

\item{output_eic}{output_EIC or not.}
}
\description{
Get the retention time of internal standards.
}
\author{
Xiaotao Shen
}
